backend/apps/api/pyproject.toml
[project]
name = "methlab-api-service"
version = "0.1.0"
description = "MethLab Facility & Monitoring API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "psycopg[binary]>=3.2.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "PyJWT>=2.8.0",
    "boto3>=1.34.0",
    "methlab-common",
]

[tool.uv.sources]
methlab-common = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
